+++ src/ilcdlib/epd/dialect/itb.py
@@ -17,7 +17,9 @@
 #  Charles Pankow Foundation, Microsoft Sustainability Fund, Interface, MKA Foundation, and others.
 #  Find out more at www.BuildingTransparency.org
 #
+from ilcdlib.dto import IlcdContactInfo, OpenEpdIlcdOrg, ValidationDto
 from ilcdlib.epd.reader import IlcdEpdReader
+from ilcdlib.type import LangDef
 
 
 class ItbIlcdXmlEpdReader(IlcdEpdReader):
@@ -27,3 +29,25 @@ class ItbIlcdXmlEpdReader(IlcdEpdReader):
     def is_known_url(cls, url: str) -> bool:
         """Return whether the URL recognized as a known Itb URL."""
         return "itb" in url.lower()
+
+    def get_third_party_verifier_email(self, validations: list[ValidationDto]) -> OpenEpdIlcdOrg | None:
+        """
+        Return first third party verifier email.
+
+        ITB contains personal info instead of organization info.
+        """
+        verifier = self.get_third_party_verifier(validations)
+        if not verifier:
+            return None
+        contact = verifier.get_contact()
+        if contact is None:
+            return None
+        return IlcdContactInfo.parse_obj(contact).email
+
+    def get_product_description(self, lang: LangDef) -> str | None:
+        """Return the product description in the given language."""
+        return self._get_localized_text(
+            self.epd_el_tree,
+            ("process:processInformation", "process:technology", "process:technologicalApplicability"),
+            lang,
+        )